WIP: actually check out PR when building it
I seriously don't understand what's going on. We are not checking out a
PR when we are building it. The worst is that all tests are passing and
all the code seems to be correct. But logs tell something else, can
anyone tell me what's wrong?

Yes, I will update/change this message before merge.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>