Set initial status metric only for first target
It roughly takes 1s to create one check status on GitHub, so if a
project expect 20 status lines, it would take at least 20s to
report these.

The problem with our initial status reporting (SLO1) was that we would
measure how long it took report every single status:

    response_time = elapsed_seconds(
        begin=self.event.created_at,
        end=task_accepted_time)

(that's measured for every target)

The problem is that the last status will indeed be set 20s after
the event was accepted.

We have two options here:
1. Change SLO1 to account for this (e.g. increase to 1 minute?)
2. Change how we measure SLO1

I went for (2) because I understand SLO1 as:

> @psss writes to our chat that Packit did not set initial statuses
> after 2 minutes of him pushing to a PR

Once Packit starts reacting to the event, everyone knows something's
going on in the background. If nothing happens (no statuses reported),
something's wrong and we should investigate.

So SLO1 prevents Petr to write to our chat that Packit did not react.
And for that, Packit only needs to set a single status.

If we wanted to be 100% complete, we should have another measure to
check that all statuses were set correctly in a reasonable amount of
time - which we have with SLO2 and 3.

Context: https://sentry.io/organizations/red-hat-it/issues/3764597999
Fixes #1728

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>